refactor: extract LoginPage to auth feature directory
