+++ src/features/auth/pages/index.ts
@@ -1,2 +1,3 @@
 export * from "./PendingApprovalPage";
 export * from "./LogoutPage";
+export * from "./LoginPage";
